Update app.js
Declared a variable to manipulate the reset button element.

Declared a variable to manipulate box elements.

Added a function named "winner" to determine the logic of who won and alert it.

Added a function named "tie" to determine the logic if the game is a tie and alert it.

Added a function named "restart" that will run when someone clicks the play again button. This will clear the board and reset the game.